refactor: moved util functions into its own folder and file
